+++ package.json
@@ -4,7 +4,8 @@
     "description": "discord.js rewrite of my discord.py bot shigure",
     "main": "index.js",
     "scripts": {
-        "test": "node index"
+        "dev": "tsx watch index.ts",
+        "start": "tsx index.ts"
     },
     "repository": {
         "type": "git",
@@ -23,7 +24,7 @@
     "homepage": "https://github.com/ecfidler/shigure-js#readme",
     "dependencies": {
         "bufferutil": "^4.0.6",
-        "discord.js": "^14.18.0",
+        "discord.js": "^14.19.1",
         "file-type": "^16.5.3",
         "node-gyp": "^9.0.0",
         "saucenao": "^0.0.2",
@@ -34,7 +35,9 @@
     "devDependencies": {
         "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.5.0",
-        "prettier": "2.7.1"
+        "prettier": "2.7.1",
+        "tsx": "^4.19.3",
+        "typescript": "^5.8.3"
     },
     "prettier": {
         "tabWidth": 4,
