+++ src/commands/Pin to lowlights.ts
